backend/pyproject.toml
[project]
name = "ex-bags-backend"
version = "0.1.0"
description = "Ex-Bags platform backend"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "boto3>=1.42.64",
    "celery>=5.6.2",
    "django>=5.2.12",
    "django-cors-headers>=4.9.0",
    "django-environ>=0.13.0",
    "django-storages>=1.14.6",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "pillow>=12.1.1",
    "psycopg[binary]>=3.3.3",
    "redis>=7.3.0",
    "requests>=2.32.5",
    "smsaero-api>=3.1.1",
    "weasyprint>=68.1",
]

[dependency-groups]
dev = [
    "black>=26.3.0",
    "pytest>=9.0.2",
    "pytest-django>=4.12.0",
    "ruff>=0.15.5",
]
